fix some ubsan errors

+++ access.c
@@ -4,11 +4,11 @@
 #include "mach.h"
 
 int
-get4(Map *map, ulong addr, int space, long *x)
+get4(Map *map, ulong addr, int space, ulong *x)
 {
     // get4 calls a lot of nested functions, simply to read 4 bytes.
     // - lofty
     USED(space);
-    *x = map->data[addr+3] | (map->data[addr+2] << 8) | (map->data[addr+1] << 16) | (map->data[addr] << 24);
+    *x = map->data[addr+3] | (map->data[addr+2] << 8) | (map->data[addr+1] << 16) | ((ulong)(map->data[addr]) << 24);
     return 1;
 }
